Restructure B45.
